Conflicts:
	PowerEditor/src/ScintillaComponent/AutoCompletion.cpp
	PowerEditor/src/ScintillaComponent/AutoCompletion.h
	PowerEditor/src/ScintillaComponent/Buffer.cpp
	PowerEditor/src/ScintillaComponent/Buffer.h
	PowerEditor/src/ScintillaComponent/DocTabView.cpp
	PowerEditor/src/ScintillaComponent/DocTabView.h
	PowerEditor/src/ScintillaComponent/FindReplaceDlg.cpp
	PowerEditor/src/ScintillaComponent/FindReplaceDlg.h
	PowerEditor/src/ScintillaComponent/FunctionCallTip.cpp
	PowerEditor/src/ScintillaComponent/FunctionCallTip.h
	PowerEditor/src/ScintillaComponent/GoToLineDlg.cpp
	PowerEditor/src/ScintillaComponent/GoToLineDlg.h
	PowerEditor/src/ScintillaComponent/Printer.cpp
	PowerEditor/src/ScintillaComponent/Printer.h
	PowerEditor/src/ScintillaComponent/ScintillaCtrls.cpp
	PowerEditor/src/ScintillaComponent/ScintillaEditView.cpp
	PowerEditor/src/ScintillaComponent/ScintillaEditView.h
	PowerEditor/src/ScintillaComponent/SmartHighlighter.cpp
	PowerEditor/src/ScintillaComponent/SmartHighlighter.h
	PowerEditor/src/ScintillaComponent/UserDefineDialog.cpp
	PowerEditor/src/ScintillaComponent/UserDefineDialog.h
	PowerEditor/src/ScintillaComponent/colors.h
	PowerEditor/src/ScintillaComponent/columnEditor.cpp
	PowerEditor/src/ScintillaComponent/columnEditor.h
	PowerEditor/src/ScintillaComponent/xmlMatchedTagsHighlighter.cpp
	PowerEditor/src/ScintillaComponent/xmlMatchedTagsHighlighter.h
	PowerEditor/src/WinControls/DockingWnd/DockingDlgInterface.h

+++ PowerEditor/src/ScintillaComponent/DocTabView.h
@@ -71,9 +71,9 @@ public :
 	BufferID findBufferByName(const TCHAR * fullfilename);	//-1 if not found, something else otherwise
 
 	int getIndexByBuffer(BufferID id);
-	BufferID getBufferByIndex(int index);
+	BufferID getBufferByIndex(size_t index);
 
-	void setBuffer(int index, BufferID id);
+	void setBuffer(size_t index, BufferID id);
 
 	static bool setHideTabBarStatus(bool hideOrNot) {
 		bool temp = _hideTabBarStatus;
